requirements.txt
aiosmtpd==1.4.6
alembic==1.13.1
atpublic==6.0.1
attrs==25.3.0
Babel==2.14.0
blinker==1.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
elastic-transport==8.17.1
elasticsearch==9.0.2
email_validator==2.1.1
Flask==3.0.3
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
guess_language-spirit==0.5.3
idna==3.6
iniconfig==2.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
langdetect==1.0.9
Mako==1.3.2
MarkupSafe==2.1.5
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.8.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2024.1
redis==6.2.0
requests==2.31.0
rq==2.4.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
urllib3==2.2.1
visitor==0.1.3
Werkzeug==3.0.2
WTForms==3.1.2
